requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.0.0
